Consolidate metrics/events_database into sync index & recorded_event files
