+++ glean/src/core/metrics/events_database/index.ts
@@ -2,30 +2,26 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-import type SynchronousStore from "../../storage/sync.js";
-import type { Event } from "./recorded_event.js";
+import type Store from "../../storage.js";
+import type { Event, ExtraValues } from "./recorded_event.js";
 import type { JSONArray, JSONValue } from "../../utils.js";
-import type MetricsDatabaseSync from "../database/sync.js";
-import type ErrorManagerSync from "../../error/sync.js";
-import type { InternalEventMetricType as EventMetricType } from "../types/event.js";
 
-import { isUndefined } from "../../utils.js";
-import log, { LoggingLevel } from "../../log.js";
 import { Context } from "../../context.js";
-import { RecordedEvent } from "./recorded_event.js";
 import {
   EVENTS_PING_NAME,
   GLEAN_EXECUTION_COUNTER_EXTRA_KEY,
   GLEAN_REFERENCE_TIME_EXTRA_KEY
 } from "../../constants.js";
 import { ErrorType } from "../../error/error_type.js";
-import {
-  createDateObject,
-  EVENT_DATABASE_LOG_TAG,
-  getExecutionCounterMetric,
-  getGleanRestartedEventMetric,
-  removeTrailingRestartedEvents
-} from "./shared.js";
+import { Lifetime } from "../lifetime.js";
+import log, { LoggingLevel } from "../../log.js";
+import { InternalEventMetricType as EventMetricType } from "../types/event.js";
+import { InternalCounterMetricType as CounterMetricType } from "../types/counter.js";
+import { RecordedEvent } from "./recorded_event.js";
+import { generateReservedMetricIdentifiers } from "../database.js";
+import { isString, isUndefined } from "../../utils.js";
+
+const EVENT_DATABASE_LOG_TAG = "core.Metric.EventsDatabase";
 
 /// UTILS ///
 /**
@@ -46,13 +42,122 @@ function recordGleanRestartedEvent(sendInPings: string[], time = Context.startTi
   );
 }
 
-// See `IEventsDatabase` for method documentation.
-export class EventsDatabaseSync {
-  private eventsStore: SynchronousStore;
+/**
+ * Attempts to create a date object from a string.
+ *
+ * Throws if unsuccessful.
+ *
+ * @param str The string to generate a date from.
+ * @returns The Date object created.
+ */
+export function createDateObject(str?: ExtraValues): Date {
+  if (!isString(str)) {
+    str = "";
+  }
+
+  const date = new Date(str);
+  // Date object will not throw errors when we attempt to create an invalid date.
+  if (isNaN(date.getTime())) {
+    throw new Error(`Error attempting to generate Date object from string: ${str}`);
+  }
+  return date;
+}
+
+/**
+ * Creates an execution counter metric.
+ *
+ * @param sendInPings The list of pings this metric is sent in.
+ *        Note: The 'events' ping should not contain glean.restarted events,
+ *        so this ping will be filtered out from the 'sendInPings' array.
+ * @returns A metric type instance.
+ */
+export function getExecutionCounterMetric(sendInPings: string[]): CounterMetricType {
+  return new CounterMetricType({
+    ...generateReservedMetricIdentifiers("execution_counter"),
+    sendInPings: sendInPings.filter((name) => name !== EVENTS_PING_NAME),
+    lifetime: Lifetime.Ping,
+    disabled: false
+  });
+}
+
+/**
+ * Creates an `glean.restarted` event metric.
+ *
+ * @param sendInPings The list of pings this metric is sent in.
+ *        Note: The 'events' ping should not contain glean.restarted events,
+ *        so this ping will be filtered out from the 'sendInPings' array.
+ * @returns A metric type instance.
+ */
+export function getGleanRestartedEventMetric(sendInPings: string[]): EventMetricType {
+  return new EventMetricType(
+    {
+      category: "glean",
+      name: "restarted",
+      sendInPings: sendInPings.filter((name) => name !== EVENTS_PING_NAME),
+      lifetime: Lifetime.Ping,
+      disabled: false
+    },
+    [GLEAN_REFERENCE_TIME_EXTRA_KEY]
+  );
+}
+
+/**
+ * Checks if the given event is a `glean.restarted` event.
+ *
+ * @param event The event to check.
+ * @returns True if the event is a `glean.restarted` event, false otherwise.
+ */
+export function isRestartedEvent(event: RecordedEvent): boolean {
+  // This extra key will only exist for `glean.restarted` events. If at some
+  // point that is no longer the case, this can be updated to check the event
+  // `name` and `category` instead.
+  return !!event?.get()?.extra?.[GLEAN_REFERENCE_TIME_EXTRA_KEY];
+}
+
+/**
+ * Removes all trailing `glean.restarted` events. We will continue to check
+ * the last element of the array until there are no longer any elements OR
+ * the event is not a `glean.restarted` event.
+ *
+ * @param sortedEvents Before this is called, events should already be sorted
+ *        which includes removing the leading `glean.restarted` event.
+ * @returns The input array without any trailing `glean.restarted` events.
+ */
+export function removeTrailingRestartedEvents(sortedEvents: RecordedEvent[]): RecordedEvent[] {
+  while (!!sortedEvents.length && isRestartedEvent(sortedEvents[sortedEvents.length - 1])) {
+    sortedEvents.pop();
+  }
+
+  return sortedEvents;
+}
+
+/**
+ * The events database is an abstraction layer on top of the underlying storage.
+ *
+ * Event data is saved to the database in the following format:
+ *
+ * {
+ *  "pingName": {
+ *    [
+ *      {
+ *        "timestamp": 0,
+ *        "category": "something",
+ *        "name": "other",
+ *        "extra": {...}
+ *      },
+ *      ...
+ *    ]
+ *  }
+ * }
+ *
+ * Events only support `Ping` lifetime.
+ */
+export class EventsDatabase {
+  private eventsStore: Store;
   private initialized = false;
 
   constructor() {
-    this.eventsStore = new Context.platform.Storage("events") as SynchronousStore;
+    this.eventsStore = new Context.platform.Storage("events");
   }
 
   /// PUBLIC ///
@@ -65,7 +170,7 @@ export class EventsDatabaseSync {
     // Submit the events ping in case there are _any_ events unsubmitted from the previous run
     if (storeNames.includes(EVENTS_PING_NAME)) {
       const storedEvents = (this.eventsStore.get([EVENTS_PING_NAME]) as JSONArray) ?? [];
-      if (storedEvents.length > 0) {
+      if (storedEvents.length > 0 && storedEvents.length >= Context.config.maxEvents) {
         Context.corePings.events.submit("startup");
       }
     }
@@ -87,7 +192,7 @@ export class EventsDatabaseSync {
       const executionCounter = getExecutionCounterMetric([ping]);
 
       let currentExecutionCount = (
-        Context.metricsDatabase as MetricsDatabaseSync
+        Context.metricsDatabase
       ).getMetric<number>(ping, executionCounter);
       // There might not be an execution counter stored in case:
       //
@@ -269,7 +374,7 @@ export class EventsDatabaseSync {
           // we increase the previous timestamp by one to make sure
           // timestamps keep increasing.
           restartedOffset = previousEventTimestamp + 1;
-          (Context.errorManager as ErrorManagerSync).record(
+          Context.errorManager.record(
             getGleanRestartedEventMetric([pingName]),
             ErrorType.InvalidValue,
             `Invalid time offset between application sessions found for ping "${pingName}". Ignoring.`
@@ -312,4 +417,4 @@ export class EventsDatabaseSync {
   }
 }
 
-export default EventsDatabaseSync;
+export default EventsDatabase;
